tests/wpt/tests/css/cssom-view/screen-detached-frame.html

<!doctype html>
<meta charset="utf-8">
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-window-screen">
<link rel="help" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1858977">
<link rel="author" href="mailto:emilio@crisal.io" title="Emilio Cobos Ãlvarez">
<link rel="author" href="https://mozilla.org" title="Mozilla">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe></iframe>
<script>
onload = function() {
  test(() => {
    let frame = document.querySelector("iframe");
    let win = frame.contentWindow;
    frame.remove();
    assert_true(!!win.screen, "Window.screen should be available");
    for (let prop of ["top", "left", "width", "height"]) {
      let availProp = "avail" + prop[0].toUpperCase() + prop.substr(1);
      if (prop == "width" || prop == "height") {
        assert_true(prop in win.screen, prop + "must be implemented per spec")
        assert_true(availProp in win.screen, availProp + "must be implemented per spec")
      }
      if (prop in win.screen) {
        assert_equals(win.screen[prop], 0, prop);
      }
      if (availProp in win.screen) {
        assert_equals(win.screen[availProp], 0, availProp);
      }
    }

    // https://drafts.csswg.org/cssom-view/#dom-screen-colordepth
    //   If the user agent does not know the color depth or does not want to
    //   return it for privacy considerations, it should return 24.
    for (let prop of ["colorDepth", "pixelDepth"]) {
      assert_equals(win.screen[prop], 24, prop);
    }
  }, "Window.screen on detached frame");
};
</script>
