third_party/blink/web_tests/external/wpt/css/cssom-view/screen-detached-frame-expected.txt

This is a testharness.js-based test.
[FAIL] Window.screen on detached frame
  assert_equals: colorDepth expected 24 but got 0
Harness: the test ran to completion.
